requirements.txt
# =========================
# Web / API
# =========================
fastapi
uvicorn

# =========================
# Environment
# =========================
python-dotenv

# =========================
# HTTP / Scraping
# =========================
requests
urllib3
scrapy
itemadapter
beautifulsoup4
unidecode

# =========================
# Data / Parsing
# =========================
pandas
numpy
pdfplumber
pdf2image
Pillow

# =========================
# NLP / ML / Embeddings
# =========================
torch
sentence-transformers
transformers
scikit-learn
nltk
sentencepiece
accelerate
huggingface_hub

# =========================
# AI APIs
# =========================
google-generativeai

# =========================
# Database
# =========================
asyncpg
psycopg2-binary
pgvector
supabase
